добавляет подключение точки входа и модулей
